+++ CHANGELOG.md
@@ -4,7 +4,11 @@ Instead, edit the files in dev/changes/, then run 'make docs' to update this fil
 -->
 # Changelog - Go Build Action
 
-## Unreleased Changes (targeting v0.1.6)
+## Unreleased Changes (targeting v0.1.7)
+
+
+
+## [v0.1.6](https://github.com/hashicorp/actions-go-build/releases/tag/v0.1.6) - October 19, 2022
 
 Changed: